Changes translates method to make use of nested_attributes  / autosave instead of custom methods
This version makes use of built-in ActiveRecord functionality instead of
reinventing the wheel with the globalize stash method. This makes the
integration with other plugins like paper_trail easier, as you do not
have to create a custom solution for globalize.

A lot of tests (see lost functionality) were removed), all other tests
were adjusted as needed

Lost functionality:
* setting the locale with :locale => :de within update_attributes
* translation versions
* fallbacks
* migrations
* paper_trail integration (which did not work)